fixed min charge for max kilometer delivery

+++ src/main/java/com/tabaldi/api/payload/ShippingCostPayload.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ShippingCostPayload {
 
-    private double distance;
+    private int distance;
     private double shippingCost;
     private long vendorId;
 }
